+++ app/models/event.py
@@ -21,8 +21,8 @@ class Event(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(40), nullable=False)
     address = db.Column(db.String(40), nullable=False)
-    lat = db.Column(db.Integer, nullable=False)
-    lng = db.Column(db.Integer, nullable=False)
+    lat = db.Column(db.Float, nullable=False)
+    lng = db.Column(db.Float, nullable=False)
     #TODO delete unnecessary rows with old address components 
     # city = db.Column(db.String(40), nullable=False)
     # state = db.Column(db.String(40), nullable=False)
